Bunch of schema/class renames
We now have an oracle which is the ultimate question answering technology.
