+++ ts/packages/agents/spelunker/src/selectorSchema.ts
